etc-merge: Refactor and fix dir perms while merging
While merging, existing directory in new_etc was being recursively
deleted which is not correct as any new files might also be deleted.

Instead, we simply create a directory if it doesn't exists, or if it
does exists, we update its metadata accordingly.

Add some test cases for the above.

Signed-off-by: Johan-Liebert1 <pragyanpoudyal41999@gmail.com>

cli: Add internal opt for printing etc-diff

Signed-off-by: Johan-Liebert1 <pragyanpoudyal41999@gmail.com>