composefs-backend: Update the order of deletion
Delete the boot entries first, the image second and everything else
afterwards. If we fail to delete the boot entry, then there's no point
in deleting the image as the boot entry will still show, but there will
be no image.

We delete the objects at the end, as when we later perform a gc
operation and don't find the image that references these objects, we can
remove them then.

The state directory shouldn't have any effect on boot if the image
associated to it doesn't exist.

If the staged file /run/composefs/staged-deployment does exist, but we
have already deleted the staged image, the finalize service would fail
but that wouldn't break anything

Signed-off-by: Pragyan Poudyal <pragyanpoudyal41999@gmail.com>