composefs/bls: Properly reuse kernel + initrd duplicates
Function `find_vmlinuz_initrd_duplicates` returns all the deployment ids
that match the current deployment's kernel + initrd SHA256 sum, but only
would actually exist on disk, every other deployment's BLS Config would
simply be pointing to one single directory.

We were incorrectly picking the first entry out of the returned list,
which might or might not exist as an actual directory containing the
kernel.

Fix the bug by iterating over all directory entries in `/boot` or
`/EFI/Linux` and checking if the deployment id in the list actually
exists on the disk or not

Signed-off-by: Pragyan Poudyal <pragyanpoudyal41999@gmail.com>